Extract generators to separate crate
