+++ crates/bitwarden-generators/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "bitwarden-generators"
+version = "0.1.0"
+edition = "2021"
+
+
+[features]
+mobile = ["uniffi"] # Mobile-specific features
+
+[dependencies]
+bitwarden-crypto = { path = "../bitwarden-crypto", version = "=0.1.0" }
+rand = ">=0.8.5, <0.9"
+reqwest = { version = ">=0.11, <0.12", features = [
+    "json",
+], default-features = false }
+schemars = { version = ">=0.8.9, <0.9", features = ["uuid1", "chrono"] }
+serde = { version = ">=1.0, <2.0", features = ["derive"] }
+serde_json = ">=1.0.96, <2.0"
+thiserror = ">=1.0.40, <2.0"
+uniffi = { version = "=0.25.2", optional = true }
+
+[dev-dependencies]
+rand_chacha = "0.3.1"
+tokio = { version = "1.35.1", features = ["rt", "macros"] }
+wiremock = "0.5.22"
