+++ crates/bitwarden-generators/src/username_forwarders/firefox.rs
@@ -3,13 +3,13 @@ use reqwest::{
     StatusCode,
 };
 
-use crate::error::Result;
+use crate::GeneratorError;
 
 pub async fn generate(
     http: &reqwest::Client,
     api_token: String,
     website: Option<String>,
-) -> Result<String> {
+) -> Result<String, GeneratorError> {
     generate_with_api_url(http, api_token, website, "https://relay.firefox.com".into()).await
 }
 
@@ -18,7 +18,7 @@ async fn generate_with_api_url(
     api_token: String,
     website: Option<String>,
     api_url: String,
-) -> Result<String> {
+) -> Result<String, GeneratorError> {
     #[derive(serde::Serialize)]
     struct Request {
         enabled: bool,
@@ -41,7 +41,7 @@ async fn generate_with_api_url(
         .await?;
 
     if response.status() == StatusCode::UNAUTHORIZED {
-        return Err("Invalid Firefox Relay API key".into());
+        return Err(GeneratorError::InvalidApiKey);
     }
 
     // Throw any other errors
@@ -60,24 +60,30 @@ async fn generate_with_api_url(
 mod tests {
     use serde_json::json;
 
+    use crate::GeneratorError;
+
     #[tokio::test]
     async fn test_mock_success() {
         use wiremock::{matchers, Mock, ResponseTemplate};
 
-        let (server, _client) =
-            crate::util::start_mock(vec![Mock::given(matchers::path("/api/v1/relayaddresses/"))
-                .and(matchers::method("POST"))
-                .and(matchers::header("Content-Type", "application/json"))
-                .and(matchers::header("Authorization", "Token MY_TOKEN"))
-                .and(matchers::body_json(json!({
-                    "enabled": true,
-                    "generated_for": "example.com",
-                    "description": "example.com - Generated by Bitwarden."
-                })))
-                .respond_with(ResponseTemplate::new(201).set_body_json(json!({
-                    "full_address": "ofuj4d4qw@mozmail.com"
-                })))
-                .expect(1)])
+        let server = wiremock::MockServer::start().await;
+
+        server
+            .register(
+                Mock::given(matchers::path("/api/v1/relayaddresses/"))
+                    .and(matchers::method("POST"))
+                    .and(matchers::header("Content-Type", "application/json"))
+                    .and(matchers::header("Authorization", "Token MY_TOKEN"))
+                    .and(matchers::body_json(json!({
+                        "enabled": true,
+                        "generated_for": "example.com",
+                        "description": "example.com - Generated by Bitwarden."
+                    })))
+                    .respond_with(ResponseTemplate::new(201).set_body_json(json!({
+                        "full_address": "ofuj4d4qw@mozmail.com"
+                    })))
+                    .expect(1),
+            )
             .await;
 
         let address = super::generate_with_api_url(
@@ -97,19 +103,23 @@ mod tests {
     async fn test_mock_without_website() {
         use wiremock::{matchers, Mock, ResponseTemplate};
 
-        let (server, _client) =
-            crate::util::start_mock(vec![Mock::given(matchers::path("/api/v1/relayaddresses/"))
-                .and(matchers::method("POST"))
-                .and(matchers::header("Content-Type", "application/json"))
-                .and(matchers::header("Authorization", "Token MY_OTHER_TOKEN"))
-                .and(matchers::body_json(json!({
-                    "enabled": true,
-                    "description": "Generated by Bitwarden."
-                })))
-                .respond_with(ResponseTemplate::new(201).set_body_json(json!({
-                    "full_address": "856f7765@mozmail.com"
-                })))
-                .expect(1)])
+        let server = wiremock::MockServer::start().await;
+
+        server
+            .register(
+                Mock::given(matchers::path("/api/v1/relayaddresses/"))
+                    .and(matchers::method("POST"))
+                    .and(matchers::header("Content-Type", "application/json"))
+                    .and(matchers::header("Authorization", "Token MY_OTHER_TOKEN"))
+                    .and(matchers::body_json(json!({
+                        "enabled": true,
+                        "description": "Generated by Bitwarden."
+                    })))
+                    .respond_with(ResponseTemplate::new(201).set_body_json(json!({
+                        "full_address": "856f7765@mozmail.com"
+                    })))
+                    .expect(1),
+            )
             .await;
 
         let address = super::generate_with_api_url(
@@ -129,18 +139,22 @@ mod tests {
     async fn test_mock_invalid_token() {
         use wiremock::{matchers, Mock, ResponseTemplate};
 
-        let (server, _client) =
-            crate::util::start_mock(vec![Mock::given(matchers::path("/api/v1/relayaddresses/"))
-                .and(matchers::method("POST"))
-                .and(matchers::header("Content-Type", "application/json"))
-                .and(matchers::header("Authorization", "Token MY_FAKE_TOKEN"))
-                .and(matchers::body_json(json!({
-                    "enabled": true,
-                    "generated_for": "example.com",
-                    "description": "example.com - Generated by Bitwarden."
-                })))
-                .respond_with(ResponseTemplate::new(401))
-                .expect(1)])
+        let server = wiremock::MockServer::start().await;
+
+        server
+            .register(
+                Mock::given(matchers::path("/api/v1/relayaddresses/"))
+                    .and(matchers::method("POST"))
+                    .and(matchers::header("Content-Type", "application/json"))
+                    .and(matchers::header("Authorization", "Token MY_FAKE_TOKEN"))
+                    .and(matchers::body_json(json!({
+                        "enabled": true,
+                        "generated_for": "example.com",
+                        "description": "example.com - Generated by Bitwarden."
+                    })))
+                    .respond_with(ResponseTemplate::new(401))
+                    .expect(1),
+            )
             .await;
 
         let error = super::generate_with_api_url(
@@ -152,7 +166,7 @@ mod tests {
         .await
         .unwrap_err();
 
-        assert!(error.to_string().contains("Invalid Firefox Relay API key"));
+        assert_eq!(error.to_string(), GeneratorError::InvalidApiKey.to_string());
 
         server.verify().await;
     }
