+++ crates/bitwarden-generators/src/username_forwarders/simplelogin.rs
@@ -1,12 +1,12 @@
 use reqwest::{header::CONTENT_TYPE, StatusCode};
 
-use crate::error::Result;
+use crate::GeneratorError;
 
 pub async fn generate(
     http: &reqwest::Client,
     api_key: String,
     website: Option<String>,
-) -> Result<String> {
+) -> Result<String, GeneratorError> {
     generate_with_api_url(http, api_key, website, "https://app.simplelogin.io".into()).await
 }
 
@@ -15,7 +15,7 @@ async fn generate_with_api_url(
     api_key: String,
     website: Option<String>,
     api_url: String,
-) -> Result<String> {
+) -> Result<String, GeneratorError> {
     let query = website
         .as_ref()
         .map(|w| format!("?hostname={}", w))
@@ -37,7 +37,7 @@ async fn generate_with_api_url(
         .await?;
 
     if response.status() == StatusCode::UNAUTHORIZED {
-        return Err("Invalid SimpleLogin API key.".into());
+        return Err(GeneratorError::InvalidApiKey);
     }
 
     // Throw any other errors
@@ -55,37 +55,46 @@ async fn generate_with_api_url(
 #[cfg(test)]
 mod tests {
     use serde_json::json;
+
+    use crate::GeneratorError;
     #[tokio::test]
     async fn test_mock_server() {
         use wiremock::{matchers, Mock, ResponseTemplate};
 
-        let (server, _client) = crate::util::start_mock(vec![
-            // Mock the request to the SimpleLogin API, and verify that the correct request is made
-            Mock::given(matchers::path("/api/alias/random/new"))
-                .and(matchers::method("POST"))
-                .and(matchers::query_param("hostname", "example.com"))
-                .and(matchers::header("Content-Type", "application/json"))
-                .and(matchers::header("Authentication", "MY_TOKEN"))
-                .and(matchers::body_json(json!({
-                    "note": "Website: example.com. Generated by Bitwarden."
-                })))
-                .respond_with(ResponseTemplate::new(201).set_body_json(json!({
-                    "alias": "simplelogin.yut3g8@aleeas.com",
-                })))
-                .expect(1),
-            // Mock an invalid token request
-            Mock::given(matchers::path("/api/alias/random/new"))
-                .and(matchers::method("POST"))
-                .and(matchers::query_param("hostname", "example.com"))
-                .and(matchers::header("Content-Type", "application/json"))
-                .and(matchers::header("Authentication", "MY_FAKE_TOKEN"))
-                .and(matchers::body_json(json!({
-                    "note": "Website: example.com. Generated by Bitwarden."
-                })))
-                .respond_with(ResponseTemplate::new(401))
-                .expect(1),
-        ])
-        .await;
+        let server = wiremock::MockServer::start().await;
+
+        // Mock the request to the SimpleLogin API, and verify that the correct request is made
+        server
+            .register(
+                Mock::given(matchers::path("/api/alias/random/new"))
+                    .and(matchers::method("POST"))
+                    .and(matchers::query_param("hostname", "example.com"))
+                    .and(matchers::header("Content-Type", "application/json"))
+                    .and(matchers::header("Authentication", "MY_TOKEN"))
+                    .and(matchers::body_json(json!({
+                        "note": "Website: example.com. Generated by Bitwarden."
+                    })))
+                    .respond_with(ResponseTemplate::new(201).set_body_json(json!({
+                        "alias": "simplelogin.yut3g8@aleeas.com",
+                    })))
+                    .expect(1),
+            )
+            .await;
+        // Mock an invalid token request
+        server
+            .register(
+                Mock::given(matchers::path("/api/alias/random/new"))
+                    .and(matchers::method("POST"))
+                    .and(matchers::query_param("hostname", "example.com"))
+                    .and(matchers::header("Content-Type", "application/json"))
+                    .and(matchers::header("Authentication", "MY_FAKE_TOKEN"))
+                    .and(matchers::body_json(json!({
+                        "note": "Website: example.com. Generated by Bitwarden."
+                    })))
+                    .respond_with(ResponseTemplate::new(401))
+                    .expect(1),
+            )
+            .await;
 
         let address = super::generate_with_api_url(
             &reqwest::Client::new(),
@@ -105,9 +114,11 @@ mod tests {
         )
         .await
         .unwrap_err();
-        assert!(fake_token_error
-            .to_string()
-            .contains("Invalid SimpleLogin API key."));
+
+        assert_eq!(
+            fake_token_error.to_string(),
+            GeneratorError::InvalidApiKey.to_string()
+        );
 
         server.verify().await;
     }
