+++ crates/bitwarden/Cargo.toml
@@ -21,6 +21,7 @@ mobile = [
     "uniffi",
     "internal",
     "bitwarden-crypto/mobile",
+    "bitwarden-generators/mobile",
 ] # Mobile-specific features
 
 wasm-bindgen = ["chrono/wasmbind"]
@@ -30,6 +31,7 @@ base64 = ">=0.21.2, <0.22"
 bitwarden-api-api = { path = "../bitwarden-api-api", version = "=0.2.3" }
 bitwarden-api-identity = { path = "../bitwarden-api-identity", version = "=0.2.3" }
 bitwarden-crypto = { path = "../bitwarden-crypto", version = "=0.1.0" }
+bitwarden-generators = { path = "../bitwarden-generators", version = "0.1.0" }
 chrono = { version = ">=0.4.26, <0.5", features = [
     "clock",
     "serde",
