+++ crates/bitwarden/src/lib.rs
@@ -75,3 +75,9 @@ pub use client::Client;
 // Ensure the readme docs compile
 #[doc = include_str!("../README.md")]
 mod readme {}
+
+pub mod generators {
+    pub use bitwarden_generators::{
+        PassphraseGeneratorRequest, PasswordGeneratorRequest, UsernameGeneratorRequest,
+    };
+}
