--- crates/bitwarden/src/tool/generators/mod.rs
@@ -1,10 +1,0 @@
-mod client_generator;
-mod passphrase;
-mod password;
-mod username;
-mod username_forwarders;
-
-pub use client_generator::ClientGenerator;
-pub use passphrase::PassphraseGeneratorRequest;
-pub use password::PasswordGeneratorRequest;
-pub use username::{AppendType, ForwarderServiceType, UsernameGeneratorRequest};
