+++ crates/bitwarden/src/tool/mod.rs
@@ -1,8 +1,4 @@
 mod exporters;
-mod generators;
-
 pub use exporters::{ClientExporters, ExportFormat};
-pub use generators::{
-    AppendType, ClientGenerator, ForwarderServiceType, PassphraseGeneratorRequest,
-    PasswordGeneratorRequest, UsernameGeneratorRequest,
-};
+mod client_generator;
+pub use client_generator::ClientGenerator;
